+++ src/pages/ChangePasswordPage.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import ChangePassword from '../components/ChangePassword';
+import ChangePassword from '../components/MyPage/ChangePassword';
 
 function ChangePasswordPage() {
   return (
